+++ src/main.ts
@@ -26,8 +26,6 @@ const isVerbose = flags.includes("--verbose");
 const wantsToUpdate = flags.includes("--update");
 const wantsMaxim = flags.includes("--maxim");
 
-const statsModular = flags.includes("--full");
-
 async function init(update?: boolean) {
     await CheckForUpdates(update);
     await FreshSetup();
@@ -48,7 +46,7 @@ if (ParseFlag("experimental-stats", false).some((flag) => flags.includes(flag)))
     await init();
     if (confirm("Experimental features are hidden behind a flag for a reason. Are you sure?")) {
         await LogStuff(`Okay, there we go. Report any ${I_LIKE_JS.FKN} error you find in GitHub.`);
-        await TheStatistics(statsModular);
+        await TheStatistics();
     }
     Deno.exit(0);
 }